Bumped up to 2020.2
The bulk of the change came from the patch provided in the issue
comment. We build with `bazel build //ijwb:ijwb_bazel_zip --define=ij_product=intellij-2020.1`
so the change that needs to be made to WORKSPACE is a little different.

So far I've just tested that the plugin loads successfully. Will need
the next few days to build more confidence that this change actually
works.

See-Also: https://github.com/bazelbuild/intellij/issues/1567#issuecomment-653918548