Quick hack to work around the problem
Referenced ScalaTestRunLineMarkerProvider appears to exist in
scalaUltimate.jar when I look at that in IDE, but it fails to compile
with 'bazel build //ijwb:ijwb_bazel_zip --define=ij_product=intellij-canary'

I'm not entirely convinced if my IDE is correctly, so it's possible that
IDE is still pointing to an older version of scala plugin. And perhaps
if so this code still exists in a newer version of a plugin somewhere
else, but I can't find it.

So just as a quick hack I'm removing a reference to this class. We don't
use Scala anyway.